move error.ts to auth
