+++ packages/lucia-auth/src/auth/error.ts
