+++ packages/lucia-auth/src/index.ts
@@ -1,5 +1,5 @@
 export { lucia as default, SESSION_COOKIE_NAME } from "./auth/index.js";
-export { LuciaError } from "./error.js";
+export { LuciaError } from "./auth/error.js";
 export { generateRandomString } from "./utils/crypto.js";
 export { serializeCookie } from "./utils/cookie.js";
 export * from "./types.js";
